+++ app/routes/register/admin_register.py
@@ -1,5 +1,5 @@
 from flask import Blueprint, render_template, redirect, url_for, flash, request
-from ..forms import AdminRegisterForm
+from ..utils.forms import AdminRegisterForm
 from app.models.database import get_cursor, close_db_connection
 from ..utils.utils import hash_password, verify_recaptcha
 import mysql.connector
