+++ app/routes/register/user_register.py
@@ -1,5 +1,5 @@
 from flask import Blueprint, render_template, redirect, url_for, flash, request
-from ..forms import UserRegisterForm
+from ..utils.forms import UserRegisterForm
 from app.models.database import get_cursor, close_db_connection
 from ..utils.utils import hash_password, verify_recaptcha
 import mysql.connector
